feat: prevents triggering USER_EVENT on every mode change
The mode can change *a lot* in a short period of time (several times in
a second when doing a surround with mini.ai for instance). The issue is
that we were calling `enable` each time the mode changed, which
triggered the `USER_EVENT` events and in particular the
`apply_virtual_texts` function. This led to a lot of unnecessary
computation and redraw of the diagnostics, in turn making Neovim laggy,
especially in the presence of many diagnostics.

We instead only fire the event if we were disabled. If the text or the
position of the cursor has changed, the other events
(`DiagnosticChanged`, `CursorMoved`, ..) should take care of updating
the diagnostics and the virtual text.